+++ TweetProjectServer/src/isep/tweetproject/server/Tweet.java
@@ -6,15 +6,15 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement
 public class Tweet {
-	private int id;
-	private int authorId;
+	private long id;
+	private long authorId;
 	private String message;
 	private Date date;
 
 	public Tweet() {
 	}
 
-	public Tweet(int id, int authorId, String message, Date date) {
+	public Tweet(long id, long authorId, String message, Date date) {
 		this.id = id;
 		this.authorId = authorId;
 		this.message = message;
@@ -26,19 +26,19 @@ public class Tweet {
 				+ date + " : " + message;
 	}
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 
-	public int getAuthorId() {
+	public long getAuthorId() {
 		return authorId;
 	}
 
-	public void setAuthorId(int authorId) {
+	public void setAuthorId(long authorId) {
 		this.authorId = authorId;
 	}
