+++ TweetProjectServer/src/isep/tweetproject/server/User.java
@@ -6,7 +6,7 @@ import javax.xml.bind.annotation.XmlRootElement;
 
 @XmlRootElement
 public class User {
-	private int id;
+	private long id;
 	private String name;
 	private String twitterNickname;
 	private Date joinedDate;
@@ -14,7 +14,7 @@ public class User {
 	public User(){
 		
 	}
-	public User(int id, String name, String nickname, Date d)
+	public User(long id, String name, String nickname, Date d)
 	{
 		this.id = id;
 		this.name = name;
@@ -28,10 +28,10 @@ public class User {
 				+ name + " joined : " + joinedDate;
 	}
 	
-	public int getId() {
+	public long getId() {
 		return id;
 	}
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
 	public String getName() {
